+++ atlasdb-cassandra/src/main/java/com/palantir/atlasdb/cassandra/CassandraKeyValueServiceRuntimeConfig.java
@@ -90,6 +90,15 @@ public abstract class CassandraKeyValueServiceRuntimeConfig implements KeyValueS
         return 6;
     }
 
+    /**
+     * Setting this value to true will cause us to take a more conservative approach to retrying requests on exceptions.
+     * @return
+     */
+    @Value.Default
+    public boolean conservativeRequestExceptionHandler() {
+        return false;
+    }
+
     public static CassandraKeyValueServiceRuntimeConfig getDefault() {
         return ImmutableCassandraKeyValueServiceRuntimeConfig.builder().build();
     }
